Use single class for DeliveredReceiver
